Add lead name to test case failure
On 2015-08-17 I was noticing that the failure report[1] had some issues with over 30+ repeated test failures
but many of them were all on the form:

<jira>.nofixversion.nobody

Making it hard to easily see who should be reacting to these since component lead was missing from the overview.

I've thus added the component lead to the issue title so it should show up more clear.

This will though mean most likely that the history count for all will be reset to 0 on the first run; but
going forward we should be able to see it more clearly.

[1] http://jenkins.mw.lab.eng.bos.redhat.com/hudson/job/jbosstools-jiralint-weekly/lastCompletedBuild/testReport/